# lib/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rulxy Platform API",
    "description": "Internal API documentation for the Rulxy platform. Covers project management, code generation, builds, proxy services, billing, and admin endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase JWT token"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RUX-API-Key",
        "description": "RUX API key for proxy services"
      }
    }
  },
...
  "paths": {
...
    "/api/creds/connect-google": {
      "post": {
        "summary": "Connect Google Play credentials",
        "description": "Saves encrypted Google Play service account credentials for EAS builds. Requires authentication and a Pro or Elite plan. Overwrites existing credentials with the same name.",
        "tags": ["Credentials"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "serviceAccountJson", "packageName"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name for the credential"
                  },
                  "serviceAccountJson": {
                    "type": "string",
                    "description": "Google service account JSON as string"
                  },
                  "packageName": {
                    "type": "string",
                    "description": "Android package name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credentials saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error or invalid JSON" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Plan does not allow credentials" },
          "404": { "description": "User not found" },
          "500": { "description": "Failed to save credentials" }
        }
      },
      "get": {
        "summary": "Check Google credentials status",
        "description": "Retrieves the list of connected Google Play credentials for the authenticated user.",
        "tags": ["Credentials"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Credentials retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "connected": { "type": "boolean" },
                    "credentials": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "verified": { "type": "boolean" },
                          "packageName": { "type": "string" },
                          "createdAt": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Failed to check credentials" }
        }
      },
      "delete": {
        "summary": "Remove Google credentials",
        "description": "Deletes a specific Google Play credential by ID. Requires the credential to belong to the authenticated user.",
        "tags": ["Credentials"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID of the credential to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Credential deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "description": "Credential ID required" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Failed to delete credentials" }
        }
      }
    },
...
    "/api/blog/{slug}": {
      "get": {
        "summary": "Get a published blog post by slug",
        "description": "Public endpoint returning a single published blog post.",
        "tags": ["Blog"],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Post retrieved" },
          "404": { "description": "Post not found" },
          "500": { "description": "Failed to fetch post" }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List all users (admin only)",
        "description": "Returns a paginated list of all users with their details.",
        "tags": ["Admin"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          { "in": "query", "name": "search", "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "plan",
            "schema": { "type": "string", "enum": ["FREE", "PRO", "ELITE"] }
          }
        ],
        "responses": {
          "200": { "description": "Users retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin access required" }
        }
      },
      "patch": {
        "summary": "Update a user (admin only)",
        "description": "Update a user's plan, role, or credits.",
        "tags": ["Admin"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "plan": {
                    "type": "string",
                    "enum": ["FREE", "PRO", "ELITE"]
                  },
                  "role": { "type": "string", "enum": ["USER", "ADMIN"] },
                  "credits": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin access required" }
        }
      }
    },
...
  }
}